effects: fan-out joins the operations more than one host implements
`All`, `all`, `allOk` and `both` move to `effects/common`, with
`effects/node` re-exporting them — the same live coupling as `Sandbox`,
`Catch` and `Import`, since `NodeOp` unions `All` and call sites name it
through that module.

The layering argument was already recorded and already sufficient: fan-out
is an interpreter's job rather than a walk's, so a host that has
concurrency implements it and a host that does not answers the effects in
turn, with the shared logic reading the same either way.

What is new is the second implementer the rule prefers. A browser page fans
out its module loads, and a `Promise.all` in its interpreter is as much an
implementation of this operation as Node's is. That is the next step's
reason for wanting it: keeping the page's 141 loads parallel is the
interpreter's business, not something a pure walk should quietly
sequentialise.

`okList` moves with `allOk`, being its half. Three imports left unused in
`effects/node` by the move are removed with it.
